+++ js/lui.js
@@ -3,7 +3,7 @@
 	angular.module('moment', []).factory('moment', function () { return window.moment; });
 	angular.module('underscore', []).factory('_', function () { return window._; });
 
-	angular.module('lui.directives', ['pascalprecht.translate', 'moment', 'underscore', 'ui.select', 'ui.bootstrap', 'lui.filters']);
+	angular.module('lui.directives', ['pascalprecht.translate', 'moment', 'underscore', 'ui.select', 'ui.bootstrap', "ngImgCrop", 'lui.filters']);
 	angular.module('lui.filters', ['moment']);
 	angular.module('lui.services', ['cgNotify']);
 
@@ -18,7 +18,8 @@
 	angular.module('lui.translates.userpicker', []);
 	angular.module('lui.translates.daterangepicker', []);
 	angular.module('lui.translates.tablegrid', []);
-	angular.module('lui.translates', ['pascalprecht.translate','lui.translates.userpicker','lui.translates.daterangepicker','lui.translates.tablegrid']);
+	angular.module('lui.translates.imagepicker', []);
+	angular.module('lui.translates', ['pascalprecht.translate','lui.translates.userpicker','lui.translates.daterangepicker','lui.translates.tablegrid', 'lui.translates.imagepicker']);
 
 	angular.module('lui', ['lui.directives','lui.services','lui.filters','lui.templates','lui.translates']);