+++ ts/image-picker/image-picker.translates.ts
@@ -1,0 +1,29 @@
+module Lui.Directive {
+	"use strict";
+	/* tslint:disable */
+	angular.module("lui.translates.imagepicker").config(["$translateProvider", function ($translateProvider) {
+		$translateProvider.translations("en", {
+			"LUIIMGPICKER_UPLOAD_IMAGE": "change picture",
+			"LUIIMGCROPPER_CROP": "Crop",
+			"LUIIMGCROPPER_DO_NOT_CROP": "Do not crop",
+		});
+		$translateProvider.translations("de", {
+
+		});
+		$translateProvider.translations("es", {
+
+		});
+		$translateProvider.translations("fr", {
+			"LUIIMGPICKER_UPLOAD_IMAGE": "changer l'image",
+			"LUIIMGCROPPER_CROP": "Recadrer",
+			"LUIIMGCROPPER_DO_NOT_CROP": "Ne pas recadrer",
+		});
+		$translateProvider.translations("it", {
+
+		});
+		$translateProvider.translations("nl", {
+
+		});
+	}]);
+	/* tslint:enable */
+}